✨ feat: api 연동 테스트
